requirements.txt
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
DictObject==1.1.1
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.1
libmagic==1.0
luckydonald-utils==0.83
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.25.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.1.2
platformdirs==3.6.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PySocks==1.7.1
pytgbot==5.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
pyzmq==25.1.0
redis==4.5.5
requests==2.31.0
selenium==4.10.0
setuptools==67.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
spotipy==2.23.0
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
webdriver-manager==3.8.6
wheel==0.40.0
wsproto==1.2.0
zipp==3.15.0
